BREAKING CHANGE: refactor api into namespaces, swagger docs generation
