+++ templates/zone.html.j2
@@ -90,7 +90,7 @@
                 </thead>
                 <tbody>
                     {% for record in records | sort(attribute=record_sort, reverse=(record_sort_order=='asc') ) %}
-                    {% set record_url_path = url_for("record_resource", **{"record_name": record.name, "zone_name": zone.name}) %}
+                    {% set record_url_path = url_for("records_specific_dns_record", **{"record_name": record.name, "zone_name": zone.name}) %}
                     <tr data-record-name="{{ record.name }}" data-url="{{ record_url_path }}">
                         <td><input type="checkbox"></td>
                         <td class="editable" data-field="name">{{ record.name }}</td>
@@ -120,7 +120,7 @@
                     </tr>
                     {% endfor %}
                     <!-- New Record Row -->
-                    {% set new_record_url = url_for("record_resource", **{"record_name": "_new", "zone_name": zone.name}) %}
+                    {% set new_record_url = url_for("records_dns_record", **{"name": "_new", "zone_name": zone.name}) %}
                     <tr class="new-record" data-url="{{ new_record_url }}">
                         <td><input type="checkbox" disabled></td>
                         <td class="editable" data-field="name">
@@ -129,8 +129,8 @@
                         <td class="editable" data-field="type">
                             <select>
                                 <option selected disabled value="">Select Record Type</option>
-                                {% for record_type_name, record_type_data in record_types.items() %}
-                                    <option value="{{ record_type_name }}">{{ record_type_name }}</option>
+                                {% for record_type in record_types %}
+                                    <option value="{{ record_type.type }}">{{ record_type.type }}</option>
                                 {% endfor %}
                             </select>
                         </td>
